Android 4 partial FTS compatibility

+++ app/src/main/java/fr/gaulupeau/apps/Poche/data/dao/FtsDao.java
@@ -31,23 +31,34 @@ public class FtsDao {
             "article_content_after_delete_tr"
     };
 
+    public static boolean isFtsSupported() {
+        // https://www.sqlite.org/changes.html#version_3_7_9 is required for the 'content' option
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN; // https://stackoverflow.com/a/4377116
+    }
+
     public static String getQueryString() {
         return "select " + COLUMN_ID + " from " + TABLE_NAME + " where " + TABLE_NAME + " match ";
     }
 
     public static void createAll(Database db, boolean ifNotExists) {
+        if (!isFtsSupported()) return;
+
         createViewForFts(db, ifNotExists);
         createTable(db, ifNotExists);
         createTriggers(db, ifNotExists);
     }
 
     public static void dropAll(Database db, boolean ifExists) {
+        if (!isFtsSupported()) return;
+
         dropTriggers(db, ifExists);
         dropTable(db, ifExists);
         dropViewForFts(db, ifExists);
     }
 
     public static void deleteAllArticles(Database db) {
+        if (!isFtsSupported()) return;
+
         dropTable(db, true);
         createTable(db, true);
     }
